+++ src/mo_frontend/static.ml
@@ -11,75 +11,56 @@ allow re-exporting names in modules.
 *)
 
 open Mo_def
+
 open Source
 open Syntax
 
-type env = { msg_store : Diag.msg_store;
-             allow_var : bool;
-             allow_sys_call : bool;
-             allow_include : bool;
-           }
-
-let err env at =
+let err m at =
   let open Diag in
-  add_msg env.msg_store
-   (error_message
-      at
-      "M0014"
-      "type"
-      (if not (env.allow_sys_call) then
-         "non-static expression in library, module or migration expression"
-       else
-         "non-static expression in actor body compiled with enhanced migration capabilities."))
+  add_msg m
+    (error_message
+       at
+       "M0014"
+       "type"
+       "non-static expression in library, module or migration expression")
 
-let pat_err env at =
+let pat_err m at =
   let open Diag in
-  add_msg env.msg_store
+  add_msg m
     (error_message
        at
        "M0015"
        "type"
        "only trivial patterns allowed in static expressions")
 
-let rec exp env e  = match e.it with
+let rec exp m e = match e.it with
   (* Plain values *)
-  | HoleE (s, e) -> exp env !e
+  | HoleE (s, e) -> exp m !e
   | (PrimE _ | LitE _ | ActorUrlE _ | FuncE _) -> ()
-  | (TagE (_, exp1) | OptE exp1) -> exp env exp1
-  | TupE es -> List.iter (exp env) es
+  | (TagE (_, exp1) | OptE exp1) -> exp m exp1
+  | TupE es -> List.iter (exp m) es
   | ArrayE (mut, es) ->
     begin
       match mut.it with
-      | Const ->  List.iter (exp env) es
-      | Var -> err env e.at
+      | Const ->  List.iter (exp m) es
+      | Var -> err m e.at
     end
   | ObjBlockE (eo, _, _, dfs) ->
-    Option.iter (exp env) eo; dec_fields env dfs
+    Option.iter (exp m) eo; dec_fields m dfs
   | ObjE (bases, efs) ->
-    List.iter (exp env) bases; exp_fields env efs
+    List.iter (exp m) bases; exp_fields m efs
 
   (* Variable access. Dangerous, due to loops. *)
   | (VarE _ | ImportE _ | ImplicitLibE _) -> ()
 
   (* Projections. These are a form of evaluation. *)
   | ProjE (exp1, _)
-  | DotE (exp1, _, _) -> exp env exp1
-  | IdxE (exp1, exp2) -> err env e.at
+  | DotE (exp1, _, _) -> exp m exp1
+  | IdxE (exp1, exp2) -> err m e.at
 
   (* Transparent *)
-  | AnnotE (exp1, _) | IgnoreE exp1 | DoOptE exp1 -> exp env exp1
-  | BlockE ds -> List.iter (dec env) ds
-
-  (*
-     if <system> and we want to allow <system> calls, check.
-     use-case: multi-migration actor bodies where we want to allow e.g., timers.
-  *)
-  | CallE (_, callee, inst, (_, ref_args))
-    -> (match (env.allow_sys_call, inst.it) with
-        | (true, Some(true, _)) ->
-          (exp env callee;
-           exp env !ref_args)
-        | _ -> err env e.at)
+  | AnnotE (exp1, _) | IgnoreE exp1 | DoOptE exp1 -> exp m exp1
+  | BlockE ds -> List.iter (dec m) ds
 
   (* Clearly non-static *)
   | UnE _
@@ -97,6 +78,7 @@ let rec exp env e  = match e.it with
   | BinE _
   | RelE _
   | AssignE _
+  | CallE _
   | AndE _
   | OrE _
   | WhileE _
@@ -108,39 +90,29 @@ let rec exp env e  = match e.it with
   | TryE _
   | BangE _
   | NullCoalesceE _
-  -> err env e.at
+  -> err m e.at
 
-and dec_fields env dfs = List.iter (fun df -> dec env df.it.dec) dfs
+and dec_fields m dfs = List.iter (fun df -> dec m df.it.dec) dfs
 
-and exp_fields env efs = List.iter (fun (ef : exp_field) ->
-  if ef.it.mut.it = Var then err env ef.at;
-  exp env ef.it.exp) efs
+and exp_fields m efs = List.iter (fun (ef : exp_field) ->
+  if ef.it.mut.it = Var then err m ef.at;
+  exp m ef.it.exp) efs
 
-and dec env d = match d.it with
+and dec m d = match d.it with
   | TypD _ | ClassD _ | MixinD _ -> ()
-  | IncludeD _ ->
-    if env.allow_include
-    then ()
-    else err env d.at
-  | ExpD e -> exp env e
-  | LetD (p, e, fail) ->
-    pat env p;
-    exp env e;
-    Option.iter (exp env) fail
-  | VarD (_, e) ->
-    if env.allow_var
-    then exp env e
-    else err env d.at
+  | ExpD e -> exp m e
+  | LetD (p, e, fail) -> pat m p; exp m e; Option.iter (exp m) fail
+  | VarD _ | IncludeD _ -> err m d.at
 
-and pat env p = match p.it with
+and pat m p = match p.it with
   | (WildP | VarP _) -> ()
 
   (*
   If we allow projections above, then we should allow irrefutable
   patterns here.
   *)
-  | TupP ps -> List.iter (pat env) ps
-  | ObjP fs -> List.iter (pat_field env) fs
+  | TupP ps -> List.iter (pat m) ps
+  | ObjP fs -> List.iter (pat_field m) fs
 
   (* TODO:
     claudio: what about singleton variant patterns? These are irrefutable too.
@@ -149,37 +121,11 @@ and pat env p = match p.it with
   *)
 
   (* Everything else is forbidden *)
-  | _ -> pat_err env p.at
+  | _ -> pat_err m p.at
 
-and pat_field env pf = match pf.it with
-  | ValPF(_, p) -> pat env p
+and pat_field m pf = match pf.it with
+  | ValPF(_, p) -> pat m p
   | TypPF(_) -> ()
 
-let module_fields msg_store =
-  dec_fields {
-    msg_store;
-    allow_var = false;
-    allow_sys_call = false;
-    allow_include = false }
-let exp msg_store =
-  exp
-    { msg_store;
-      allow_var = false;
-      allow_sys_call = false;
-      allow_include = false }
-
-let actor_fields msg_store =
-  dec_fields {
-    msg_store;
-    allow_var = true;
-    allow_sys_call = true;
-    allow_include = true }
-
 let prog p =
-  Diag.with_message_store (fun msg_store ->
-      Some (List.iter (dec
-       {msg_store;
-        allow_var = false;
-        allow_sys_call = false;
-        allow_include = false })
-       p.it))
+  Diag.with_message_store (fun m -> List.iter (dec m) p.it; Some ())