+++ src/mo_frontend/static.mli
@@ -1,6 +1,5 @@
 open Mo_def
 
-val exp :  Diag.msg_store -> Syntax.exp -> unit
-val module_fields : Diag.msg_store -> Syntax.dec_field list -> unit
-val actor_fields : Diag.msg_store -> Syntax.dec_field list -> unit
-val prog :  Syntax.prog -> unit Diag.result
+val exp : Diag.msg_store -> Syntax.exp -> unit
+val dec_fields : Diag.msg_store -> Syntax.dec_field list -> unit
+val prog : Syntax.prog -> unit Diag.result